fix: 수익률이 제대로 계산이 되지 않았던 오류수정

+++ src/main/java/lotto/domain/LottoResult.java
@@ -22,8 +22,8 @@ public class LottoResult {
                 .sum();
     }
 
-    public Long calculateRevenueRate(int payment) {
-        return getTotalPrize() / payment * 100;
+    public Double calculateRevenueRate(int payment) {
+        return ((double) getTotalPrize() / payment) * 100;
     }
 
 }
